Move query creation logic out of SearchLocation
SearchLocation was doing WAY to much stuff. This is another pass to start the
process of decomposing this component into some sensible sized pieces.
