+++ package.json
@@ -10,6 +10,7 @@
     "compile": "lingui compile",
     "add-locale": "lingui add-locale",
     "dev": "npm run clean && cross-env NODE_ENV=development babel-watch server/index.js",
+    "dbg": "node --inspect-brk node_modules/jest/bin/jest.js",
     "start": "npm run serve",
     "serve": "cross-env NODE_ENV=production node buildServer/index.js",
     "build": "npm run build:client && npm run build:server && npm run build:node",
@@ -36,7 +37,7 @@
     "emotion-theming": "^9.0.0",
     "express": "^4.15.2",
     "facepaint": "^1.1.2",
-    "graphql": "^0.11.7",
+    "graphql": "^0.13.2",
     "graphql-tag": "^2.5.0",
     "history": "^4.6.3",
     "isomorphic-fetch": "^2.2.1",
